requirements.txt
pandas
numpy
matplotlib
Pillow
pyprind
scikit-learn
scipy
jupyter
kaggle-cli
tqdm
imagehash
dictionaries